[test] 테스트 코드 패키지화 및 구매 가격 예외 테스트 추가
